+++ commands/stop.js
@@ -14,16 +14,16 @@ module.exports = {
 
     let songFromDB = await Queue.find().exec()
     let songDB = songFromDB[0]
-    console.log(`--stop ${songDB.title}`)
-    msg.channel.send(`stop ${songDB.title}`)
-
-    Queue.deleteMany({}).exec(function (err, post) {
-      if (err) throw err
-      // console.log(`--stop all song`)
-      // msg.channel.send(`stop all song`)
-    })
-
-    if (serverQueue === undefined) {
+    console.log('stop com is ', songDB.length !== 0)
+    if (songDB.length !== 0) {
+      console.log(`--stop ${songDB.title}`)
+      msg.channel.send(`stop ${songDB.title}`)
+      Queue.deleteMany({}).exec(function (err, post) {
+        if (err) throw err
+        // console.log(`--stop all song`)
+        // msg.channel.send(`stop all song`)
+      })
+    } else if (serverQueue === undefined) {
       return msg.channel.send('no song to stop')
     }
     // console.log(`--stop ${serverQueue.songs[0].title}`)
